Ersha_Ecosystem_Backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
psycopg[binary]==3.2.9
django-cors-headers==4.7.0
Pillow==11.3.0
python-decouple==3.8
django-filter==25.1
requests==2.31.0
drf-yasg==1.21.7
setuptools>=80.0.0
asgiref==3.9.1
sqlparse==0.5.3
aiohttp==3.9.1
PyJWT==2.8.0
cryptography==41.0.7
google-generativeai==0.3.2
reportlab==4.0.7

# WebSocket support
channels==4.1.0
channels-redis==4.2.0
redis==5.0.1
